# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/orders/dine": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new DineOrder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DineOrders"
                ],
                "summary": "Create a new DineOrder",
                "parameters": [
                    {
                        "description": "DineOrder data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models_dine.AddDineOrderData"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/{restaurant_id}/menus/{menu_id}/categories": {
            "post": {
                "description": "Create a new category for a specific menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Category"
                ],
                "summary": "Create a new category for a specific menu",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddMenuCategoryData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/{restaurant_id}/menus/{menu_id}/categories/{category_id}/items": {
            "get": {
                "description": "Retrieve all items for a specific category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Retrieve all items for a specific category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create a new menu item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Category Item"
                ],
                "summary": "Create a new menu item",
                "parameters": [
                    {
                        "description": "Item data",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddMenuItemData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.AddMenuCategoryData": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.AddMenuItemData": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "is_vegetarian": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AddMenuItemOptionData"
                    }
                }
            }
        },
        "models.AddMenuItemOptionData": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "models_dine.AddDineOrderData": {
            "type": "object",
            "required": [
                "duration",
                "plan_id",
                "restaurant_id"
            ],
            "properties": {
                "duration": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "promo_code": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
...
    }
}